# docs/openapi/api-docs.yaml
openapi: 3.0.1
info:
  title: Spring REST Books API
  description: |
    A comprehensive REST API for managing a bookstore system including customers,
    books, authors, carts, orders, payments and shipments.

    ## Features
    - Customers.
    - Books.
    - Authors.
    - Carts.
    - Orders.
    - Payments.
    - Shipments.
  contact:
    email: carlosarroyoam@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1-SNAPSHOT
servers:
  - url: http://localhost:8081
    description: Localhost server url
...
paths:
  /customers:
    get:
      tags:
        - Customers
      summary: Gets the list of customers
      operationId: findAllCustomers
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 25
        - name: firstName
          in: query
          required: false
          schema:
            type: string
            maxLength: 128
        - name: lastName
          in: query
          required: false
          schema:
            type: string
            maxLength: 128
        - name: email
          in: query
          required: false
          schema:
            type: string
            maxLength: 128
        - name: username
          in: query
          required: false
          schema:
            type: string
            maxLength: 128
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
    post:
      tags:
        - Customers
      summary: Creates a new customer
      operationId: createCustomer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequestDto"
        required: true
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
  /customers/{customerId}:
    get:
      tags:
        - Customers
      summary: Gets a customer by its id
      operationId: findCustomerById
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
    put:
      tags:
        - Customers
      summary: Updates a customer by its id
      operationId: updateCustomer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerRequestDto"
        required: true
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
    delete:
      tags:
        - Customers
      summary: Deletes a customer by its id
      operationId: deleteCustomer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
...
  /carts:
    get:
      tags:
        - Carts
      summary: Gets the customer cart
      operationId: findCart
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
  /carts/items:
    put:
      tags:
        - Carts
      summary: Updates a cart item
      operationId: updateCartItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemRequestDto"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
  /carts/items/{cartItemId}:
    delete:
      tags:
        - Carts
      summary: Deletes a cart item
      operationId: deleteCartItem
      parameters:
        - name: cartItemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
...
components:
  schemas:
    CreateCustomerRequestDto:
      required:
        - first_name
        - last_name
        - username
        - email
      type: object
      properties:
        first_name:
          maxLength: 128
          minLength: 3
          type: string
        last_name:
          maxLength: 128
          minLength: 3
          type: string
        email:
          maxLength: 128
          minLength: 3
          type: string
        username:
          maxLength: 128
          minLength: 3
          type: string
    UpdateCustomerRequestDto:
      required:
        - first_name
        - last_name
      type: object
      properties:
        first_name:
          maxLength: 128
          minLength: 3
          type: string
        last_name:
          maxLength: 128
          minLength: 3
          type: string
...
    UpdateCartItemRequestDto:
      required:
        - book_id
        - shipping_address
        - billing_address
        - items
      type: object
      properties:
        book_id:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
          minimum: 1
...
    CustomerDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        username:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CartDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItemDto"
        customer_id:
          type: number
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CartItemDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        quantity:
          type: number
        book_id:
          type: number
          format: int64
        added_at:
          type: string
          format: date-time
...
    AppExceptionDto:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
        status:
          type: integer
        details:
          type: object
          additionalProperties:
            type: string
        path:
          type: string
        timestamp:
          format: date-time
  responses:
    Created:
      description: Created
    NoContent:
      description: No content
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppExceptionDto"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppExceptionDto"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppExceptionDto"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppExceptionDto"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppExceptionDto"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT