# src/main/java/com/carlosarroyoam/rest/books/cart/dto/CartDto.java
package com.carlosarroyoam.rest.books.cart.dto;

import com.carlosarroyoam.rest.books.cart.entity.Cart;
import com.carlosarroyoam.rest.books.customer.dto.CustomerDto;
import java.time.LocalDateTime;
import java.util.List;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.mapstruct.Mapper;
import org.mapstruct.NullValuePropertyMappingStrategy;
import org.mapstruct.ReportingPolicy;
import org.mapstruct.factory.Mappers;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class CartDto {
  private Long id;
  private List<CartItemDto> items;
  private CustomerDto customer;
  private LocalDateTime createdAt;
  private LocalDateTime updatedAt;

  @Mapper(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE, unmappedTargetPolicy = ReportingPolicy.IGNORE)
  public interface CartDtoMapper {
    CartDtoMapper INSTANCE = Mappers.getMapper(CartDtoMapper.class);

    CartDto toDto(Cart entity);
  }
}